# kestra/openapi.yml
openapi: 3.0.1
info:
  title: Kestra
  license:
    name: Apache 2.0
    url: https://raw.githubusercontent.com/kestra-io/kestra/master/LICENSE
  version: 1.0.0
...
paths:
...
  /api/v1/roles:
    get:
      tags:
      - Roles
      summary: List roles
      operationId: list_2
      responses:
        "200":
          description: list_2 200 response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Role"
    post:
      tags:
      - Roles
      summary: Create role
      operationId: create_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleController.CreateRoleRequest"
        required: true
      responses:
        "200":
          description: create_1 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
  /api/v1/roles/{id}:
    get:
      tags:
      - Roles
      summary: Get role by id
      operationId: get_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: get_1 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
    delete:
      tags:
      - Roles
      summary: Delete role
      operationId: delete_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: delete_1 200 response
    patch:
      tags:
      - Roles
      summary: Update role
      operationId: update
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleController.UpdateRoleRequest"
        required: true
      responses:
        "200":
          description: update 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
...
  /api/v1/{tenant}/namespaces/{namespace}/variables:
    get:
      tags:
      - Namespaces
      summary: List namespace variables
      operationId: listVariables
      parameters:
      - name: namespace
        in: path
        description: The namespace id
        required: true
        schema:
          type: string
      - name: tenant
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: listVariables 200 response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties:
                    type: string
    put:
      tags:
      - Namespaces
      summary: Set namespace variables (replace all)
      operationId: setVariables
      parameters:
      - name: namespace
        in: path
        description: The namespace id
        required: true
        schema:
          type: string
      - name: tenant
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: string
        required: true
      responses:
        "200":
          description: setVariables 200 response
          content:
            application/json:
              schema:
                type: object
  /api/v1/{tenant}/namespaces/{namespace}/variables/{key}:
    delete:
      tags:
      - Namespaces
      summary: Delete a namespace variable
      operationId: deleteVariable
      parameters:
      - name: namespace
        in: path
        description: The namespace id
        required: true
        schema:
          type: string
      - name: key
        in: path
        description: The variable key
        required: true
        schema:
          type: string
      - name: tenant
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: deleteVariable 200 response
          content:
            application/json:
              schema:
                type: object
...
components:
  schemas:
...
    Role:
      required:
      - id
      - name
      - permissions
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        permissions:
          type: array
          items:
            type: string
    RoleController.CreateRoleRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        permissions:
          type: array
          items:
            type: string
    RoleController.UpdateRoleRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        permissions:
          type: array
          items:
            type: string
...